requirements.txt
fastapi
uvicorn
pandas
numpy
python-doctr
difflib 
python-multipart
